Improved/fixed the examples a bit - some still need conversion to v2
Old code was not exiting with errors if there were issues due to faulty
promise chaining so all of them have this now. The new v2 Hatch
constructor API is used everywhere also but the code for the v1
ingesters was not changed since it was pretty far out of the scope of
the original story.