Added language to hatch constructor and extra fileds in manifest
Language is needed when packing so we need to pass it in durng hatch
creation and we now also put the hatch name and language in the
hatch_metadata file to be easily parsed by the feed processing pipeline.

Also removed now is the requirement to pass the argv array since we can
grab it directly and only handle it if overrides are needed.